requirements.txt
# Core Framework
Django==4.2.10  # Latest LTS version of Django
djangorestframework==3.14.0  # Django REST Framework for API development
django-cors-headers==4.3.1  # Handles CORS for cross-domain requests

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter for production databases
dj-database-url==2.1.0  # Database URL configuration for deployment environments

# Authentication and Authorization
djangorestframework-simplejwt==5.3.1  # JWT authentication for API
django-allauth==0.61.0  # Authentication, registration, account management
django-oauth-toolkit==2.3.0  # OAuth2 support

# API Documentation
drf-yasg==1.21.7  # Swagger/OpenAPI documentation
markdown==3.6  # Markdown support for API docs

# Security
django-filter==24.1  # Filtering for API endpoints
django-defender==0.9.8  # Blocking and tracking login attempts
django-csp==3.7  # Content Security Policy middleware

# Performance
django-redis==5.4.0  # Redis cache backend for Django
django-storages==1.14.3  # Storage backends for static/media files (S3, Azure, etc.)
whitenoise==6.6.0  # Serve static files efficiently

# Production Server
gunicorn==21.2.0  # WSGI HTTP server for production
uvicorn==0.28.0  # ASGI server for WebSockets support

# Monitoring and Logging
sentry-sdk==1.42.0  # Error tracking and performance monitoring
django-prometheus==2.3.1  # Prometheus metrics for monitoring

# Testing
pytest==8.0.2  # Testing framework
pytest-django==4.8.0  # Django support for pytest
coverage==7.4.3  # Code coverage measurement

# Utilities
python-dotenv==1.0.1  # Environment variable management
pillow==10.2.0  # Image processing
celery==5.3.6  # Distributed task queue for background jobs
django-celery-beat==2.5.0  # Database-backed periodic tasks
django-celery-results==2.5.1  # Store Celery task results in Django database
